% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factiva.R, R/import.R, R/integrum.R,
%   R/kikuzo.R, R/lexis.R, R/nexis.R, R/yomidasu.R
\name{import_factiva}
\alias{import_factiva}
\alias{import}
\alias{import_integrum}
\alias{import_kikuzo}
\alias{import_lexis}
\alias{import_nexis}
\alias{import_yomidasu}
\title{Extract texts and meta data from newspaper databases}
\usage{
import_factiva(path, paragraph_separator = "\\n\\n")

import_integrum(path, paragraph_separator = "\\n\\n")

import_kikuzo(path, paragraph_separator = "\\n\\n")

import_lexis(path, paragraph_separator = "\\n\\n")

import_nexis(path, paragraph_separator = "\\n\\n")

import_yomidasu(path, paragraph_separator = "\\n\\n")
}
\arguments{
\item{path}{either path to a file or a directory that contain files}

\item{paragraph_separator}{a character to separate paragraphs in body
texts.}
}
\description{
These functions extract headings, body texts and meta data (date, byline, length,
section, edition etc.) from files downloaded from various newspaper databases.
}
\examples{
\dontrun{
# Factiva
iri <- import_factiva("testthat/data/factiva/irish-independence_1_2017-11-14.html")
chi <- import_factiva("testthat/data/factiva/chosun_ilbo_1_2018-05-28.html")
all <- import_factiva("testthat/data/factiva")
}


\dontrun{
# Integrum
xlsx <- import_integrum("tests/data/integrum/channel1_20050201-20050228_0001-0059_0059.xlsx")
html <- import_integrum("tests/data/integrum/moscow_1993-06-06_1994-09-02.html")
}
\dontrun{
# Kikuzo
kik <- import_kikuzo("tests/data/kikuzo/kikuzo_1985-01-01_001.html")
}
\dontrun{
# Lexis Advance
nyt <- import_lexis("tests/data/lexis/nyt1.docx")
ust <- import_lexis("tests/data/lexis/usa-today.docx")
wsp <- import_lexis("tests/data/lexis/washington-post.docx")
all <- import_lexis("tests/data/lexis/")
}
\dontrun{
# Nexis UK
irt <- import_nexis("tests/data/nexis/irish-times_1995-06-12_0001.html")
afp <- import_nexis("tests/data/nexis/afp_2013-03-12_0501.html")
gur <- import_nexis("tests/data/nexis/guardian_1986-01-01_0001.html")
spg <- import_nexis("tests/data/nexis/spiegel_2012-02-01_0001.html")
}
\dontrun{
# Yomidasu
yom <- import_yomidasu("tests/data/yomidasu/yomidasu_1987-01-01_001.html")
}
}
