% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factiva.R
\name{import_factiva}
\alias{import_factiva}
\title{Extract texts and meta data from factiva HTML files}
\usage{
import_factiva(path, paragraph_separator = "\\n\\n")
}
\arguments{
\item{path}{either path to a HTML file or a directory that containe HTML files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body texts.}
}
\description{
This extract headings, body texts and meta data (date, byline, length,
secotion, edntion) from HTML files downloaded from the Factiva database.
}
\examples{
\dontrun{
one <- import_factiva("testthat/data/factiva/irish-independence_1_2017-11-14.html")
two <- import_factiva("testthat/data/factiva/chosun_ilbo_1_2018-05-28.html")
all <- import_factiva("testthat/data/factiva")
}


}
