% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrum.R
\name{import_integrum}
\alias{import_integrum}
\title{Extract texts and meta data from Integrum XLSX files}
\usage{
import_integrum(path, paragraph_separator = "\\n\\n")
}
\arguments{
\item{path}{either path to a XLSX file or a directory that containe XLSX
files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body
texts.}
}
\description{
This extract headings, body texts and meta data (date, time) from XLSX files
downloaded from the Integrum database.
}
\examples{
\dontrun{
one <- import_integrum("tests/data/integrum/channel1_20050201-20050228_0001-0059_0059.xlsx")
two <- import_integrum("tests/data/integrum/ntv_20121201-20121231_0001-0100_0190.xlsx")
all <- import_integrum("tests/data/integrum")
}
}
