% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kikuzo.R
\name{import_kikuzo}
\alias{import_kikuzo}
\title{Extract texts and meta data from Kikuzo HTML files}
\usage{
import_kikuzo(path, paragraph_separator = "|")
}
\arguments{
\item{path}{either path to a HTML file or a directory that containe HTML files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body texts.}
}
\description{
This extract headings, body texts and meta data (date, byline, length,
secotion, edntion) from HTML files downloaded from the Kikuzo database.
}
\examples{
\dontrun{
one <- import_kikuzo("testthat/data/kikuzo/asahi_1985-01-01_001.html")
two <- import_kikuzo("testthat/data/kikuzo/asahi_1985-01-01_002.html")
all <- import_kikuzo("testthat/data/kikuzo")
}
}
