% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yomidasu.R
\name{import_yomidasu}
\alias{import_yomidasu}
\title{Extract texts and meta data from Yomidasu HTML files}
\usage{
import_yomidasu(path, paragraph_separator = "|")
}
\arguments{
\item{path}{either path to a HTML file or a directory that containe HTML files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body texts.}
}
\description{
This extract headings, body texts and meta data (date, byline, length,
secotion, edntion) from HTML files downloaded from the Yomidasu database.
}
\examples{
\dontrun{
one <- import_yomidasu("tests/data/yomidasu/yomidasu_1987-01-01_001.html")
two <- import_yomidasu("tests/data/yomidasu/yomidasu_2018-01-01_001.html")
all <- import_yomidasu("tests/data/yomidasu/")
}
}
