% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gaps}
\alias{check_gaps}
\title{Check and plot gaps in data}
\usage{
check_gaps(x, size = 7, plot = TRUE)
}
\arguments{
\item{x}{out from import functions, or any data.frame with "date" column}

\item{size}{size of gap to detect specified in number of days, defaults to 7 days}

\item{plot}{if \code{TRUE}, plot the number of items for each day}
}
\description{
Check and plot gaps in data
}
