% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexis.R
\name{import_lexis}
\alias{import_lexis}
\title{Extract texts and meta data from Lexis files}
\usage{
import_lexis(path, paragraph_separator = "\\n\\n", variant = c("advance"))
}
\arguments{
\item{path}{either path to a file or a directory that containe files}

\item{paragraph_separator}{a character to sperarate paragrahphs in body
texts.}

\item{variant}{specify type of Lexis database files downloaded. Files should be
MS Word (.docx) for Lexis Advance.}
}
\description{
This extract headings, body texts and meta data (date, byline, length,
secotion, edntion) from items in HTML or MS Word files downloaded from the
Lexis database.
}
\examples{
\dontrun{
# Lexis Advance
nyt <- import_lexis("tests/data/lexis/nyt1.docx")
ust <- import_lexis("tests/data/lexis/usa-today.docx")
wsp <- import_lexis("tests/data/lexis/washington-post.docx")
all <- import_lexis("tests/data/lexis/")
}
}
